[app] - refactoring of roots during import - (turns out if the folder has name starting with Lowercase it gives an error when importing component to another) The solution is whether to call folders with lowercase but during import use capital case or simply call folders with capitalcase
